refactor(plugin-githubactions-nodejs): adapt to new scaffold
